chore: 파일 분리, 변수명 변경, format 적용
