+++ alsongDalsong/alsongDalsong/alsongDalsong/Sources/Views/Result/HummingResultViewModel.swift
@@ -14,21 +14,21 @@ final class HummingResultViewModel: @unchecked Sendable {
     private var roomInfoRepository: RoomInfoRepositoryProtocol
     private var musicRepository: MusicRepositoryProtocol
     private var cancellables = Set<AnyCancellable>()
-    
+
     @Published var isNext: Bool = false
     @Published var currentResult: Answer?
     @Published var currentRecords: [ASEntity.Record] = []
     @Published var currentsubmit: Answer?
     @Published var recordOrder: UInt8? = 0
     @Published var isHost: Bool = false
-    
+
     // 미리 받아놓을 정보 배열
     private var recordsResult: [ASEntity.Record] = []
     private var submitsResult: Answer?
     private var roomNumber: String = ""
-    
+
     var hummingResult: [(answer: ASEntity.Answer, records: [ASEntity.Record], submit: ASEntity.Answer)] = []
-    
+
     init(hummingResultRepository: HummingResultRepositoryProtocol,
          avatarRepository: AvatarRepositoryProtocol,
          gameStatusRepository: GameStatusRepositoryProtocol,
@@ -44,8 +44,9 @@ final class HummingResultViewModel: @unchecked Sendable {
         self.roomActionRepository = roomActionRepository
         self.roomInfoRepository = roomInfoRepository
         self.musicRepository = musicRepository
+        fetchResult()
     }
-    
+
     func fetchResult() {
         hummingResultRepository.getResult()
             .receive(on: DispatchQueue.main)
@@ -54,17 +55,17 @@ final class HummingResultViewModel: @unchecked Sendable {
                 Logger.debug(completion)
             } receiveValue: { [weak self] (result: [(answer: ASEntity.Answer, records: [ASEntity.Record], submit: ASEntity.Answer, recordOrder: UInt8)]) in
                 guard let self else { return }
-                
+
                 if (result.count - 1) < result.first?.recordOrder ?? 0 { return }
                 Logger.debug("호출")
                 hummingResult = result.map {
                     (answer: $0.answer, records: $0.records, submit: $0.submit)
                 }
-                
+
                 hummingResult.sort {
                     $0.answer.player?.order ?? 0 < $1.answer.player?.order ?? 1
                 }
-                
+
                 Logger.debug("hummingResult \(hummingResult)")
 
                 let current = hummingResult.removeFirst()
@@ -73,21 +74,21 @@ final class HummingResultViewModel: @unchecked Sendable {
                 submitsResult = current.submit
             }
             .store(in: &cancellables)
-        
+
         playerRepository.isHost()
             .receive(on: DispatchQueue.main)
             .sink { [weak self] isHost in
                 self?.isHost = isHost
             }
             .store(in: &cancellables)
-        
+
         roomInfoRepository.getRoomNumber()
             .receive(on: DispatchQueue.main)
             .sink { [weak self] roomNumber in
                 self?.roomNumber = roomNumber
             }
             .store(in: &cancellables)
-        
+
         Publishers.CombineLatest(gameStatusRepository.getStatus(), gameStatusRepository.getRecordOrder())
             .receive(on: DispatchQueue.main)
             .sink { status, _ in
@@ -100,10 +101,10 @@ final class HummingResultViewModel: @unchecked Sendable {
             }
             .store(in: &cancellables)
     }
-    
+
     func startPlaying() async {
         await startPlayingCurrentMusic()
-        
+
         while !recordsResult.isEmpty {
             currentRecords.append(recordsResult.removeFirst())
             guard let fileUrl = currentRecords.last?.fileUrl else { continue }
@@ -117,14 +118,14 @@ final class HummingResultViewModel: @unchecked Sendable {
         }
         currentsubmit = submitsResult
     }
-    
-    private func startPlayingCurrentMusic() async -> Void {
+
+    private func startPlayingCurrentMusic() async {
         guard let fileUrl = currentResult?.music?.previewUrl else { return }
         let data = await musicRepository.getMusicData(url: fileUrl)
         await AudioHelper.shared.startPlaying(data, option: .partial(time: 10))
         await waitForPlaybackToFinish()
     }
-    
+
     private func waitForPlaybackToFinish() async {
         await withCheckedContinuation { continuation in
             Task {
@@ -135,7 +136,7 @@ final class HummingResultViewModel: @unchecked Sendable {
             }
         }
     }
-    
+
     func nextResultFetch() {
         if hummingResult.isEmpty { return }
         let current = hummingResult.removeFirst()
@@ -145,7 +146,7 @@ final class HummingResultViewModel: @unchecked Sendable {
         currentRecords.removeAll()
         currentsubmit = nil
     }
-    
+
     private func getRecordData(url: URL?) -> AnyPublisher<Data?, Error> {
         if let url {
             hummingResultRepository.getRecordData(url: url)
@@ -156,13 +157,13 @@ final class HummingResultViewModel: @unchecked Sendable {
                 .eraseToAnyPublisher()
         }
     }
-    
+
     private func fetchRecordData(url: URL) async throws -> Data {
         try await withCheckedThrowingContinuation { continuation in
             getRecordData(url: url)
                 .compactMap { $0 }
                 .sink(receiveCompletion: { completion in
-                    if case .failure(let error) = completion {
+                    if case let .failure(error) = completion {
                         continuation.resume(throwing: error)
                     }
                 }, receiveValue: { data in
@@ -171,12 +172,12 @@ final class HummingResultViewModel: @unchecked Sendable {
                 .store(in: &cancellables)
         }
     }
-    
+
     func getAvatarData(url: URL?) async -> Data? {
         guard let url else { return nil }
         return await avatarRepository.getAvatarData(url: url)
     }
-    
+
     func changeRecordOrder() async throws {
         do {
             try await roomActionRepository.changeRecordOrder(roomNumber: roomNumber)
@@ -185,7 +186,7 @@ final class HummingResultViewModel: @unchecked Sendable {
             throw error
         }
     }
-    
+
     func navigationToLobby() async throws {
         do {
             if !hummingResult.isEmpty { return }
@@ -194,12 +195,12 @@ final class HummingResultViewModel: @unchecked Sendable {
             throw error
         }
     }
-    
+
     func getArtworkData(url: URL?) async -> Data? {
         guard let url else { return nil }
         return await musicRepository.getMusicData(url: url)
     }
-    
+
     func cancelSubscriptions() {
         cancellables.removeAll()
     }
