+++ alsongDalsong/alsongDalsong/alsongDalsong/Sources/Views/Result/ResultList.swift
@@ -1,0 +1,104 @@
+import UIKit
+import Combine
+import ASEntity
+
+final class ResultList: UIView {
+    private let albumImageView = UIImageView()
+    private let musicNameLabel = UILabel()
+    private let singerNameLabel = UILabel()
+    private let titleLabel = UILabel()
+    private var cancellables = Set<AnyCancellable>()
+
+    init() {
+        super.init(frame: .zero)
+        setupView()
+        setupConstraints()
+    }
+
+    required init?(coder: NSCoder) {
+        super.init(coder: coder)
+        setupView()
+        setupConstraints()
+    }
+    
+    func bind(
+        to dataSource: Published<Answer?>.Publisher,
+        fetcher: @escaping (URL?) async -> Data?,
+        musicFetcher: @escaping () async -> Void?
+    ) {
+        dataSource
+            .receive(on: DispatchQueue.main)
+            .compactMap { $0 }
+            .sink { [weak self] answer in
+                self?.musicNameLabel.text = answer.music?.title
+                self?.singerNameLabel.text = answer.music?.artist
+                Task {
+                    guard let url = answer.music?.artworkUrl else { return }
+                    let data = await fetcher(url)
+                    self?.setImage(data: data)
+                    await musicFetcher()
+                }
+            }
+            .store(in: &cancellables)
+    }
+    
+    private func setImage(data: Data?) {
+        guard let data else { return }
+        albumImageView.image = UIImage(data: data)
+    }
+
+    private func setupView() {
+        backgroundColor = .asSystem
+        
+        titleLabel.text = "정답은..."
+        titleLabel.font = .font(ofSize: 24)
+        titleLabel.textColor = .asBlack
+        titleLabel.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(titleLabel)
+
+        albumImageView.contentMode = .scaleAspectFill
+        albumImageView.layer.cornerRadius = 6
+        albumImageView.clipsToBounds = true
+        albumImageView.translatesAutoresizingMaskIntoConstraints = false
+        albumImageView.image = UIImage(named: "mojojojo") // Placeholder image
+        addSubview(albumImageView)
+
+        musicNameLabel.font = .font(ofSize: 24)
+        musicNameLabel.textColor = .asBlack
+        musicNameLabel.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(musicNameLabel)
+
+        singerNameLabel.font = .font(ofSize: 24)
+        singerNameLabel.textColor = UIColor.gray
+        singerNameLabel.translatesAutoresizingMaskIntoConstraints = false
+        addSubview(singerNameLabel)
+        
+        layer.cornerRadius = 12
+        layer.shadowColor = UIColor.asShadow.cgColor
+        layer.shadowOffset = CGSize(width: 4, height: 4)
+        layer.shadowRadius = 0
+        layer.shadowOpacity = 1.0
+        layer.borderWidth = 3
+        layer.borderColor = UIColor.black.cgColor
+    }
+
+    private func setupConstraints() {
+        NSLayoutConstraint.activate([
+            titleLabel.topAnchor.constraint(equalTo: topAnchor, constant: 16),
+            titleLabel.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+
+            albumImageView.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 12),
+            albumImageView.leadingAnchor.constraint(equalTo: leadingAnchor, constant: 16),
+            albumImageView.widthAnchor.constraint(equalToConstant: 60),
+            albumImageView.heightAnchor.constraint(equalToConstant: 60),
+
+            musicNameLabel.topAnchor.constraint(equalTo: albumImageView.topAnchor),
+            musicNameLabel.leadingAnchor.constraint(equalTo: albumImageView.trailingAnchor, constant: 15),
+            musicNameLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16),
+
+            singerNameLabel.topAnchor.constraint(equalTo: musicNameLabel.bottomAnchor, constant: 4),
+            singerNameLabel.leadingAnchor.constraint(equalTo: albumImageView.trailingAnchor, constant: 15),
+            singerNameLabel.trailingAnchor.constraint(equalTo: trailingAnchor, constant: -16)
+        ])
+    }
+}
